- add support for --add-associationDeclarations - add support for --add-ecpHome - refactoring solution extend to break files between fmm and dashui
